+++ src/const/env.ts
@@ -1,4 +1,8 @@
-export const DEPLOYMENT_ENDPONTS = {
+export const DEPLOYMENT_ENDPOINTS = {
     NEW_JOB: "https://platform.ploomber.io/dashboards/",
-    REGISTER_API: "https://www.platform.ploomber.io/register/"
+}
+
+export const DOCS = {
+    GET_API_KEY: "https://docs.cloud.ploomber.io/en/latest/quickstart/apikey.html",
+    VOILA_EXAMPLES: "https://docs.cloud.ploomber.io/en/latest/examples/voila.html",
 }